Move general functions into utils dir
