+++ frontend/src/ts/test/github.ts
@@ -1,4 +1,6 @@
 import * as Loader from "../elements/loader";
+import { getRandomItem } from "../utils/arrays";
+import { cache } from "../utils/decorators";
 
 const fallbackCodeLanguage = "javascript";
 
@@ -126,52 +128,4 @@ async function sendGithubApiRequest(url: string): Promise<unknown> {
     throw Error(fileRequest.statusText);
   }
   return await fileRequest.json();
-}
-
-function getRandomItem<T>(list: T[]): T {
-  const randomIndex = Math.floor(Math.random() * list.length);
-  return list[randomIndex];
-}
-
-function cache<T extends unknown[], U>(
-  fn: (...args: T) => U,
-  options = {
-    cacheDurationMilliseconds: 30000,
-  }
-): (...args: T) => U {
-  let cacheTimestamp = Date.now();
-  const cache = new Map();
-
-  const isCacheExpired = (): boolean => {
-    if (cacheTimestamp === null) {
-      return true;
-    }
-
-    if (Date.now() > cacheTimestamp + options.cacheDurationMilliseconds) {
-      return true;
-    }
-
-    return false;
-  };
-
-  const cachedFn = (...args: T): U => {
-    if (isCacheExpired()) {
-      cache.clear();
-    }
-
-    const cacheKey = JSON.stringify(args);
-
-    if (cache.has(cacheKey)) {
-      return cache.get(cacheKey);
-    }
-
-    const result = fn(...args);
-
-    cache.set(cacheKey, result);
-    cacheTimestamp = Date.now();
-
-    return result;
-  };
-
-  return cachedFn;
 }
