+++ frontend/src/ts/utils/arrays.ts
@@ -1,0 +1,4 @@
+export function getRandomItem<T>(list: T[]): T {
+  const randomIndex = Math.floor(Math.random() * list.length);
+  return list[randomIndex];
+}
