+++ frontend/src/ts/utils/decorators/index.ts
@@ -1,0 +1,1 @@
+export * from "./cache";
